requirements.txt
anyascii==0.3.2
asgiref==3.7.2
backcall==0.2.0
backports.functools-lru-cache==1.6.5
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
comm==0.1.4
cryptography==41.0.5
cycler==0.12.1
cytoolz==0.12.0
dask==2023.10.1
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
Django==4.2.6
django-allauth==0.58.0
django-cogwheels==0.3
django-filter==22.1
django-modelcluster==6.1
django-permissionedforms==0.1
django-taggit==3.1.0
django-treebeard==4.7
djangorestframework==3.14.0
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==2.0.1
filetype==1.2.0
fsspec==2023.10.0
html5lib==1.1
idna==3.4
importlib-metadata==6.8.0
kiwisolver==1.4.4
l18n==2021.3
lazy_loader==0.3
locket==1.0.0
munkres==1.1.4
mysql-connector-python==8.2.0
mysqlclient==2.2.0
nest-asyncio==1.5.8
networkx==3.2.1
oauthlib==3.2.2
openpyxl==3.1.2
packaging==23.2
pandas==2.0.3
parso==0.8.3
partd==1.4.1
pickleshare==0.7.5
Pillow==10.0.1
pip==23.2.1
platformdirs==3.11.0
prompt-toolkit==3.0.39
protobuf==4.21.12
psutil==5.9.0
psycopg2==2.9.9
pure-eval==0.2.2
Pygments==2.16.1
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2023.3.post1
PyWavelets==1.4.1
requests==2.31.0
requests-oauthlib==1.3.1
scipy==1.11.3
seaborn==0.12.2
selenium==4.27.1
setuptools==68.0.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
telepath==0.3.1
threadpoolctl==2.2.0
tqdm==4.66.1
traitlets==5.13.0
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.11.0
tzdata==2023.3
urllib3==2.0.7
wcwidth==0.2.8
webdriver_manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
wheel==0.41.2
Willow==1.6.2
zipp==3.17.0
